EDM4hep Inputs

Extension functionality to read EDM4hep ReconstructedParticles, using the EDM4hep.jl package.

Examples

EDM4hep ReconstructedParticles can be used as direct input into jet reconstruction.

A number of working examples are maintained in the EDM4hep examples directory of the package's examples.

Here is a snippet that shows the main steps:

using EDM4hep
using EDM4hep.RootIO
using JetReconstruction

# Change this to something that works on your system
input_file = joinpath("directory", "EDM4hep.root")
reader = RootIO.Reader(input_file)
events = RootIO.get(reader, "events")

evt = events[1]

recps = RootIO.get(reader, evt, "ReconstructedParticles")

cs = jet_reconstruct(recps; algorithm = JetAlgorithm.Durham)

Function Index

JetReconstruction.energy

JetReconstruction.px

JetReconstruction.py

JetReconstruction.pz

EDM4hep Interfaces

JetReconstruction.energy — Method

JetReconstruction.energy(recoparticle::ReconstructedParticle)

Return the energy component of a ReconstructedParticle's four vector.

JetReconstruction.px — Method

JetReconstruction.px(recoparticle::ReconstructedParticle)

Return the x component of the momentum of a ReconstructedParticle.

JetReconstruction.py — Method

JetReconstruction.py(recoparticle::ReconstructedParticle)

Return the y component of the momentum of a ReconstructedParticle.

JetReconstruction.pz — Method

JetReconstruction.pz(recoparticle::ReconstructedParticle)

Return the z component of the momentum of a ReconstructedParticle.